resizer/requirements.txt
pika==1.3.1
Pillow==9.3.0
requests==2.28.1